Moved all letsencrypt config to config file, and general refactor around this area
